#81 feat-refactor: updated movegen lookup tables - Added Python scripts that generate lookup tables - Change NEIGHBOUR1 and NEIGHBOUR2 in logic::lookup to [&[CellIndex]; 45] instead of [[CellIndex]; 45*7] Some 10-20% speedup on perft, but search within margin of error

+++ src/logic/actions.rs
@@ -136,7 +136,7 @@ impl ActionTrait for Action {
 
     #[inline(always)]
     fn search_depth(self) -> u64 {
-        (self >> (3 * INDEX_WIDTH)) & 0xFF
+        ((self >> (3 * INDEX_WIDTH)) & 0xFF) as u64
     }
 
     /// Concatenate a half action and the last index into a `Action`.
